requirements.txt
fastapi==0.75.0
uvicorn[standard]~=0.17.6
jinja2~=3.1.1
APScheduler~=3.9.1
RPi.GPIO==0.7.1